Noel tch staing

See merge request autofix/autofix-app!355

+++ lib/UI/Mechanic/BottomBar/Home/upcoming_services_mdl.dart
@@ -88,7 +88,7 @@ class UpcomingCompletedService {
   int bookStatus;
   int totalPrice;
   double tax;
-  int commission;
+  double commission;
   double serviceCharge;
   dynamic totalTime;
   String serviceTime;
